+++ app/src/main/AndroidManifest.xml
@@ -13,16 +13,16 @@
         android:theme="@style/Theme.Oomok"
         tools:targetApi="31">
         <activity
-            android:name=".MainActivity"
+            android:name=".activity.MainActivity"
             android:exported="false" />
         <activity
-            android:name=".StartActivity"
+            android:name=".activity.StartActivity"
             android:exported="false" />
         <activity
-            android:name=".WinnerActivity"
+            android:name=".activity.WinnerActivity"
             android:exported="false" />
         <activity
-            android:name=".SplashActivity"
+            android:name=".activity.SplashActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
