+++ app/src/main/res/layout/activity_start.xml
@@ -4,7 +4,7 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".StartActivity">
+    tools:context=".activity.StartActivity">
 
     <LinearLayout
         android:orientation="vertical"
