+++ app/src/main/res/layout/activity_winner.xml
@@ -4,7 +4,7 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".WinnerActivity">
+    tools:context=".activity.WinnerActivity">
 
     <LinearLayout
         android:orientation="vertical"
